:bulb: 카테고리 텝바 컬랙션뷰 컴포넌트 구현
